Refactor proxy behaviors and abstractions for cleaner responsibilities
The refactoring follows the design of Unity interception which is based
on a chain of responsibility between the behaviors. The abstraction is
proven usable and sound by providing two implementation for Castle
and LinFu dynamic proxy generators. This ensures that we're not creating
an abstraction that will only serve our own static and dynamic proxy
generators :)

The new abstraction also decouples the method invocation (immutable) from
the method result or exception returns, which is mutable by the behaviors,
therefore providing better isolation and making it easier to reason about
both as different entities.

An additional "Context" was added to the invocation and passed to the method
result interfaces too, with a simple key/value store for inter-behavior data
passing if needed.

Proxies (static or dynamic) must now implement always the IProxy interface,
which allows a mocking framework (or a general-purpose AOP library) to inject
its own behavior on the proxy.

The proxy itself is kept extremely simple so that the responsibility of
complex interception behaviors is left for consuming libraries.